+++ x/gov/client/cli/prompt.go
@@ -4,14 +4,15 @@ import (
 	"encoding/json"
 	"fmt"
 	"os"
-	"reflect" // #nosec
 	"sort"
-	"strconv"
 	"strings"
 
-	"github.com/manifoldco/promptui"
+	gogoproto "github.com/cosmos/gogoproto/proto"
 	"github.com/spf13/cobra"
+	"google.golang.org/protobuf/proto"
+	"google.golang.org/protobuf/reflect/protoregistry"
 
+	"cosmossdk.io/client/v2/autocli/prompt"
 	"cosmossdk.io/core/address"
 	"cosmossdk.io/x/gov/types"
 
@@ -19,7 +20,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/client/flags"
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
+	sdkaddress "github.com/cosmos/cosmos-sdk/types/address"
 )
 
 const (
@@ -60,102 +61,6 @@ var suggestedProposalTypes = []proposalType{
 	},
 }
 
-// Prompt prompts the user for all values of the given type.
-// data is the struct to be filled
-// namePrefix is the name to be displayed as "Enter <namePrefix> <field>"
-// TODO: when bringing this in autocli, use proto message instead
-// this will simplify the get address logic
-func Prompt[T any](data T, namePrefix string, addressCodec address.Codec) (T, error) {
-	v := reflect.ValueOf(&data).Elem()
-	if v.Kind() == reflect.Interface {
-		v = reflect.ValueOf(data)
-		if v.Kind() == reflect.Ptr {
-			v = v.Elem()
-		}
-	}
-
-	for i := 0; i < v.NumField(); i++ {
-		// if the field is a struct skip or not slice of string or int then skip
-		switch v.Field(i).Kind() {
-		case reflect.Struct:
-			// TODO(@julienrbrt) in the future we can add a recursive call to Prompt
-			continue
-		case reflect.Slice:
-			if v.Field(i).Type().Elem().Kind() != reflect.String && v.Field(i).Type().Elem().Kind() != reflect.Int {
-				continue
-			}
-		}
-
-		// create prompts
-		prompt := promptui.Prompt{
-			Label:    fmt.Sprintf("Enter %s %s", namePrefix, strings.ToLower(client.CamelCaseToString(v.Type().Field(i).Name))),
-			Validate: client.ValidatePromptNotEmpty,
-		}
-
-		fieldName := strings.ToLower(v.Type().Field(i).Name)
-
-		if strings.EqualFold(fieldName, "authority") {
-			// pre-fill with gov address
-			defaultAddr, err := addressCodec.BytesToString(authtypes.NewModuleAddress(types.ModuleName))
-			if err != nil {
-				return data, err
-			}
-			prompt.Default = defaultAddr
-			prompt.Validate = client.ValidatePromptAddress
-		}
-
-		// TODO(@julienrbrt) use scalar annotation instead of dumb string name matching
-		if strings.Contains(fieldName, "addr") ||
-			strings.Contains(fieldName, "sender") ||
-			strings.Contains(fieldName, "voter") ||
-			strings.Contains(fieldName, "depositor") ||
-			strings.Contains(fieldName, "granter") ||
-			strings.Contains(fieldName, "grantee") ||
-			strings.Contains(fieldName, "recipient") {
-			prompt.Validate = client.ValidatePromptAddress
-		}
-
-		result, err := prompt.Run()
-		if err != nil {
-			return data, fmt.Errorf("failed to prompt for %s: %w", fieldName, err)
-		}
-
-		switch v.Field(i).Kind() {
-		case reflect.String:
-			v.Field(i).SetString(result)
-		case reflect.Int:
-			resultInt, err := strconv.ParseInt(result, 10, 0)
-			if err != nil {
-				return data, fmt.Errorf("invalid value for int: %w", err)
-			}
-			// If a value was successfully parsed the ranges of:
-			//      [minInt,     maxInt]
-			// are within the ranges of:
-			//      [minInt64, maxInt64]
-			// of which on 64-bit machines, which are most common,
-			// int==int64
-			v.Field(i).SetInt(resultInt)
-		case reflect.Slice:
-			switch v.Field(i).Type().Elem().Kind() {
-			case reflect.String:
-				v.Field(i).Set(reflect.ValueOf([]string{result}))
-			case reflect.Int:
-				resultInt, err := strconv.ParseInt(result, 10, 0)
-				if err != nil {
-					return data, fmt.Errorf("invalid value for int: %w", err)
-				}
-
-				v.Field(i).Set(reflect.ValueOf([]int{int(resultInt)}))
-			}
-		default:
-			// skip any other types
-			continue
-		}
-	}
-
-	return data, nil
-}
-
 type proposalType struct {
 	Name    string
 	MsgType string
@@ -163,8 +68,8 @@ type proposalType struct {
 }
 
 // Prompt the proposal type values and return the proposal and its metadata
-func (p *proposalType) Prompt(cdc codec.Codec, skipMetadata bool, addressCodec address.Codec) (*proposal, types.ProposalMetadata, error) {
-	metadata, err := PromptMetadata(skipMetadata, addressCodec)
+func (p *proposalType) Prompt(cdc codec.Codec, skipMetadata bool, addressCodec, validatorAddressCodec, consensusAddressCodec address.Codec) (*proposal, types.ProposalMetadata, error) {
+	metadata, err := PromptMetadata(skipMetadata)
 	if err != nil {
 		return nil, metadata, fmt.Errorf("failed to set proposal metadata: %w", err)
 	}
@@ -176,11 +81,7 @@ func (p *proposalType) Prompt(cdc codec.Codec, skipMetadata bool, addressCodec a
 	}
 
 	// set deposit
-	depositPrompt := promptui.Prompt{
-		Label:    "Enter proposal deposit",
-		Validate: client.ValidatePromptCoins,
-	}
-	proposal.Deposit, err = depositPrompt.Run()
+	proposal.Deposit, err = prompt.PromptString("Enter proposal deposit", ValidatePromptCoins)
 	if err != nil {
 		return nil, metadata, fmt.Errorf("failed to set proposal deposit: %w", err)
 	}
@@ -190,12 +91,35 @@ func (p *proposalType) Prompt(cdc codec.Codec, skipMetadata bool, addressCodec a
 	}
 
 	// set messages field
-	result, err := Prompt(p.Msg, "msg", addressCodec)
+	msg, err := protoregistry.GlobalTypes.FindMessageByURL(p.MsgType)
+	if err != nil {
+		return nil, metadata, fmt.Errorf("failed to find proposal msg: %w", err)
+	}
+	newMsg := msg.New()
+	govAddr := sdkaddress.Module(types.ModuleName)
+	govAddrStr, err := addressCodec.BytesToString(govAddr)
+	if err != nil {
+		return nil, metadata, fmt.Errorf("failed to convert gov address to string: %w", err)
+	}
+
+	prompt.SetDefaults(newMsg, map[string]interface{}{"authority": govAddrStr})
+	result, err := prompt.PromptMessage(addressCodec, validatorAddressCodec, consensusAddressCodec, "msg", newMsg)
 	if err != nil {
 		return nil, metadata, fmt.Errorf("failed to set proposal message: %w", err)
 	}
 
-	message, err := cdc.MarshalInterfaceJSON(result)
+	// message must be converted to gogoproto so @type is not lost
+	resultBytes, err := proto.Marshal(result.Interface())
+	if err != nil {
+		return nil, metadata, fmt.Errorf("failed to marshal proposal message: %w", err)
+	}
+
+	err = gogoproto.Unmarshal(resultBytes, p.Msg)
+	if err != nil {
+		return nil, metadata, fmt.Errorf("failed to unmarshal proposal message: %w", err)
+	}
+
+	message, err := cdc.MarshalInterfaceJSON(p.Msg)
 	if err != nil {
 		return nil, metadata, fmt.Errorf("failed to marshal proposal message: %w", err)
 	}
@@ -214,33 +138,22 @@ func getProposalSuggestions() []string {
 }
 
 // PromptMetadata prompts for proposal metadata or only title and summary if skip is true
-func PromptMetadata(skip bool, addressCodec address.Codec) (types.ProposalMetadata, error) {
+func PromptMetadata(skip bool) (types.ProposalMetadata, error) {
 	if !skip {
-		metadata, err := Prompt(types.ProposalMetadata{}, "proposal", addressCodec)
+		metadata, err := prompt.PromptStruct("proposal", types.ProposalMetadata{})
 		if err != nil {
-			return metadata, fmt.Errorf("failed to set proposal metadata: %w", err)
+			return types.ProposalMetadata{}, err
 		}
 
 		return metadata, nil
 	}
 
-	// prompt for title and summary
-	titlePrompt := promptui.Prompt{
-		Label:    "Enter proposal title",
-		Validate: client.ValidatePromptNotEmpty,
-	}
-
-	title, err := titlePrompt.Run()
+	title, err := prompt.PromptString("Enter proposal title", ValidatePromptNotEmpty)
 	if err != nil {
 		return types.ProposalMetadata{}, fmt.Errorf("failed to set proposal title: %w", err)
 	}
 
-	summaryPrompt := promptui.Prompt{
-		Label:    "Enter proposal summary",
-		Validate: client.ValidatePromptNotEmpty,
-	}
-
-	summary, err := summaryPrompt.Run()
+	summary, err := prompt.PromptString("Enter proposal summary", ValidatePromptNotEmpty)
 	if err != nil {
 		return types.ProposalMetadata{}, fmt.Errorf("failed to set proposal summary: %w", err)
 	}
@@ -262,17 +175,10 @@ func NewCmdDraftProposal() *cobra.Command {
 				return err
 			}
 
-			// prompt proposal type
-			proposalTypesPrompt := promptui.Select{
-				Label: "Select proposal type",
-				Items: getProposalSuggestions(),
-			}
-
-			_, selectedProposalType, err := proposalTypesPrompt.Run()
+			selectedProposalType, err := prompt.Select("Select proposal type", getProposalSuggestions())
 			if err != nil {
 				return fmt.Errorf("failed to prompt proposal types: %w", err)
 			}
-
 			var proposal proposalType
 			for _, p := range suggestedProposalTypes {
 				if strings.EqualFold(p.Name, selectedProposalType) {
@@ -283,17 +189,10 @@ func NewCmdDraftProposal() *cobra.Command {
 
 			// create any proposal type
 			if proposal.Name == proposalOther {
-				// prompt proposal type
-				msgPrompt := promptui.Select{
-					Label: "Select proposal message type:",
-					Items: func() []string {
-						msgs := clientCtx.InterfaceRegistry.ListImplementations(sdk.MsgInterfaceProtoName)
-						sort.Strings(msgs)
-						return msgs
-					}(),
-				}
+				msgs := clientCtx.InterfaceRegistry.ListImplementations(sdk.MsgInterfaceProtoName)
+				sort.Strings(msgs)
 
-				_, result, err := msgPrompt.Run()
+				result, err := prompt.Select("Select proposal message type:", msgs)
 				if err != nil {
 					return fmt.Errorf("failed to prompt proposal types: %w", err)
 				}
@@ -311,7 +210,7 @@ func NewCmdDraftProposal() *cobra.Command {
 
 			skipMetadataPrompt, _ := cmd.Flags().GetBool(flagSkipMetadata)
 
-			result, metadata, err := proposal.Prompt(clientCtx.Codec, skipMetadataPrompt, clientCtx.AddressCodec)
+			result, metadata, err := proposal.Prompt(clientCtx.Codec, skipMetadataPrompt, clientCtx.AddressCodec, clientCtx.ValidatorAddressCodec, clientCtx.ConsensusAddressCodec)
 			if err != nil {
 				return err
 			}